+ arreglo algunos leaks en entrenador y en mapa tmb + el mapa ahora lee correctamente retardos muy grandes (DL y dibujado) + agrego codigo para liberar memoria en mapa (cierro threaths tmb) + el entrenador borra las medallas
Nota: Queda sacar los retardos hardcodeados del mapa.

+++ PROC_MAPA/src/lib/estructurasMapa.h
@@ -19,11 +19,11 @@
 
 typedef struct
 {
-	uint16_t tiempoChequeadoDeadlock;
+	long long int tiempoChequeadoDeadlock;
 	char * batalla;
 	char * algoritmo;
 	uint16_t quantum;
-	uint16_t retardo;	//Valor en milisegundos
+	long long int retardo;	//Valor en milisegundos
 	char * ip;
 	char * puerto;
 } t_metadataMapa ;
